@@ -17,7 +17,7 @@
     "@types/uuid": "3.4.4",
     "btoa": "1.2.1",
     "cross-env": "7.0.2",
-    "date-fns": "^2.15.0",
+    "date-fns": "2.15.0",
     "lodash": "4.17.19",
     "node-fetch": "2.3.0",
     "nodemailer": "5.1.1",
@@ -29,19 +29,20 @@
   },
   "devDependencies": {
     "@types/jest": "24.0.11",
-    "eslint": "5.15.2",
-    "eslint-plugin-node": "8.0.1",
-    "eslint-plugin-typescript": "0.14.0",
+    "@typescript-eslint/eslint-plugin": "3.9.0",
+    "@typescript-eslint/parser": "3.9.0",
+    "eslint": "7.7.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-config-prettier": "6.11.0",
     "jest": "24.5.0",
-    "lint-staged": "8.1.5",
+    "lint-staged": "10.2.11",
     "nodemon": "1.18.10",
-    "prettier": "1.16.4",
+    "prettier": "2.0.5",
     "sinon": "7.2.7",
     "ts-jest": "24.0.0",
     "ts-node": "8.0.3",
     "tslint": "5.14.0",
-    "tslint-config-prettier": "1.18.0",
-    "typescript-eslint-parser": "22.0.0"
+    "tslint-config-prettier": "1.18.0"
   },
   "scripts": {
     "dev": "cross-env NODE_ENV=development nodemon --max-old-space-size=4096",
