@@ -24,8 +24,10 @@
   "devDependencies": {
     "@typescript-eslint/eslint-plugin": "3.9.0",
     "@typescript-eslint/parser": "3.9.0",
-    "eslint": "7.6.0",
+    "eslint": "7.7.0",
     "eslint-config-prettier": "6.11.0",
+    "eslint-plugin-react": "7.20.6",
+    "eslint-plugin-react-hooks": "4.0.8",
     "lint-staged": "10.2.11",
     "prettier": "2.0.5"
   },
