@@ -1,4 +1,4 @@
-import * as fs from 'fs'
+import fs from 'fs'
 
 import { isProdServer, isTestServer } from './env'
 
@@ -9,20 +9,33 @@ const privateKeyPath = '/apps/nopestradamus/backend/privkey.pem'
 
 export const getPrivateKey = () => fs.readFileSync(privateKeyPath, 'utf8')
 
-let config: any
+interface MyConfig {
+  database: {
+    host: string
+    user: string
+    database: string
+    password: string
+  }
+  port: string
+}
+
+let config: MyConfig
+
 const loadConfig = () => {
   if (fs.existsSync(dbconfigPath)) {
-    config = JSON.parse(fs.readFileSync(dbconfigPath, 'utf8'))
+    config = JSON.parse(fs.readFileSync(dbconfigPath, 'utf8')) as MyConfig
   } else {
     if (isProdServer() || isTestServer()) {
       throw new Error('config file not found')
     }
-    config = JSON.parse(fs.readFileSync(dbconfigDevPath, 'utf8'))
+    config = JSON.parse(fs.readFileSync(dbconfigDevPath, 'utf8')) as MyConfig
   }
 }
 
 export default () => {
-  if (config == null) {
+  // TODO would it be possible to use loadConfig as a type guard to make 'config' variable into defined.
+  // tslint:disable-next-line: strict-type-predicates
+  if (config === undefined) {
     loadConfig()
   }
   return config
